chg: output table update
* Improved output table

* Related work items: #411436
================== COMMIT-END ========================

+++ development/libraries/casaAPI/src/PrmLithoFraction.cpp
@@ -187,7 +187,7 @@ namespace casa
    PrmLithoFraction::PrmLithoFraction( mbapi::Model              & mdl
                                      , const std::string         & layerName
                                      , const std::vector<int>    & lithoFractionsInds
-                                     , const std::vector<size_t> & coordinates
+                                     , const std::vector<unsigned int> & coordinates
                                      )
                                      : Parameter()
                                      , m_layerName( layerName )
